r128: * schroedinger/schrodecoder.c: * schroedinger/schroencoder.c: * schroedinger/schrotables.c: * schroedinger/schrotables.h: * testsuite/tables.c: Update the quantisation tables to match the spec.  Only changes BITSTREAM slightly.

+++ schroedinger/schroedinger/schrotables.h
@@ -5,8 +5,8 @@
 #include <schroedinger/schro-stdint.h>
 #include <schroedinger/schroutils.h>
 
-extern int16_t schro_table_offset[61] SCHRO_INTERNAL;
-extern int16_t schro_table_quant[61] SCHRO_INTERNAL;
+extern uint32_t schro_table_offset[61] SCHRO_INTERNAL;
+extern uint32_t schro_table_quant[61] SCHRO_INTERNAL;
 extern uint32_t schro_table_inverse_quant[61] SCHRO_INTERNAL;
 extern uint16_t schro_table_division_factor[257] SCHRO_INTERNAL;
